Add files via upload
This is a fun game. The computer will select a random number and it will be not shown in the screen. You have to guess the number. If you guess a  Larger number than Computer's random number, then It will command you to pick a number that is lower than your first picked number. If you pick lower number then it will command you to pick higher number. If you pick the same number as the computer choose, then it will print...You win, with the number of attempted that you took to win.